feat(interface): refactor typings from monorepo modules.
