+++ packages/core/src/index.ts
@@ -2,4 +2,3 @@ export * from "./context/index";
 export * from "./factories/index";
 export * from "./programmers/index";
 export * from "./schemas/index";
-export * from "./typings/index";
