+++ packages/core/src/schemas/metadata/MetadataMap.ts
@@ -1,6 +1,9 @@
-import { IMetadataSchema, IMetadataTypeTag } from "@typia/interface";
+import {
+  ClassProperties,
+  IMetadataSchema,
+  IMetadataTypeTag,
+} from "@typia/interface";
 
-import { ClassProperties } from "../../typings/ClassProperties";
 import { MetadataSchema } from "./MetadataSchema";
 
 export class MetadataMap {
@@ -15,6 +18,7 @@ export class MetadataMap {
     this.tags = props.tags;
   }
 
+  /** @internal */
   public static create(props: ClassProperties<MetadataMap>): MetadataMap {
     return new MetadataMap(props);
   }
