--- packages/core/src/typings/index.ts
@@ -1,8 +1,0 @@
-export * from "./AssertionGuard";
-export * from "./ClassProperties";
-export * from "./Customizable";
-export * from "./OmitNever";
-export * from "./ProtobufAtomic";
-export * from "./SpecialFields";
-export * from "./ValidationPipe";
-export * from "./Writable";
