+++ packages/interface/src/typings/ClassProperties.ts
