+++ packages/interface/src/typings/OmitNever.ts
