+++ packages/interface/src/typings/ProtobufAtomic.ts
