+++ packages/interface/src/typings/index.ts
@@ -1,3 +1,4 @@
+export * from "./AssertionGuard";
 export * from "./CamelCase";
 export * from "./PascalCase";
 export * from "./Primitive";
@@ -5,3 +6,8 @@ export * from "./Resolved";
 export * from "./SnakeCase";
 
 export * from "./Atomic";
+export * from "./ClassProperties";
+export * from "./OmitNever";
+export * from "./ProtobufAtomic";
+export * from "./SpecialFields";
+export * from "./ValidationPipe";
