+++ packages/transform/src/features/json/JsonApplicationTransformer.ts
@@ -3,8 +3,7 @@ import { MetadataStorage } from "@typia/core";
 import { MetadataFactory } from "@typia/core";
 import { JsonApplicationProgrammer } from "@typia/core";
 import { MetadataSchema } from "@typia/core";
-import { ValidationPipe } from "@typia/core";
-import { IJsonSchemaApplication } from "@typia/interface";
+import { IJsonSchemaApplication, ValidationPipe } from "@typia/interface";
 import ts from "typescript";
 
 import { ITransformProps } from "../../ITransformProps";
