+++ packages/transform/src/features/json/JsonSchemaTransformer.ts
@@ -4,8 +4,7 @@ import { MetadataFactory } from "@typia/core";
 import { JsonSchemaProgrammer } from "@typia/core";
 import { JsonSchemasProgrammer } from "@typia/core";
 import { MetadataSchema } from "@typia/core";
-import { ValidationPipe } from "@typia/core";
-import { IJsonSchemaUnit } from "@typia/interface";
+import { IJsonSchemaUnit, ValidationPipe } from "@typia/interface";
 import ts from "typescript";
 
 import { ITransformProps } from "../../ITransformProps";
