+++ packages/transform/src/features/llm/LlmApplicationTransformer.ts
@@ -8,9 +8,8 @@ import {
   MetadataSchema,
   MetadataStorage,
   StatementFactory,
-  ValidationPipe,
 } from "@typia/core";
-import { ILlmApplication, ILlmSchema } from "@typia/interface";
+import { ILlmApplication, ILlmSchema, ValidationPipe } from "@typia/interface";
 import ts from "typescript";
 
 import { ITransformProps } from "../../ITransformProps";
