+++ packages/transform/src/features/llm/LlmSchemaTransformer.ts
@@ -5,8 +5,7 @@ import { MetadataFactory } from "@typia/core";
 import { LlmMetadataFactory } from "@typia/core";
 import { LlmSchemaProgrammer } from "@typia/core";
 import { MetadataSchema } from "@typia/core";
-import { ValidationPipe } from "@typia/core";
-import { ILlmSchema } from "@typia/interface";
+import { ILlmSchema, ValidationPipe } from "@typia/interface";
 import ts from "typescript";
 
 import { ITransformProps } from "../../ITransformProps";
