+++ packages/typia/src/re-exports.ts
@@ -1,5 +1,6 @@
 export {
   // validate
+  AssertionGuard,
   IValidation,
   IRandomGenerator,
   // json
