--- packages/typia/src/typings/Customizable.ts
@@ -1,5 +1,0 @@
-export type Customizable = {
-  number: number;
-  string: string;
-  bigint: bigint;
-};
