--- packages/typia/src/typings/index.ts
@@ -1,1 +1,0 @@
-export * from "./AssertionGuard";
